Moved instruction emission to backend.h/.cpp.
