+++ source/assembler/assembler_backend.h
@@ -14,24 +14,6 @@ namespace baremetal {
 		auto sign_extend(u64 x, u8 x_bits) -> u64;
 	
 		auto is_operand_of_same_kind(operand_type a, operand_type b) -> bool;
-		auto extract_operand_reg_beg(const operand& op) -> u8;
-		auto extract_operand_reg(const operand& op) -> u8;
-
-		enum mod_mode : u8 {
-			INDIRECT        = 0b00,
-			INDIRECT_DISP8  = 0b01,
-			INDIRECT_DISP32 = 0b10,
-			DIRECT          = 0b11
-		};
-
-		auto mod_rm(mod_mode mod, u8 rx, u8 rm) -> u8;
-		auto sib(u8 scale, u8 index, u8 base) -> u8;
-		auto rex(bool w, u8 rx, u8 base, u8 index) -> u8;
-
-		auto direct(u8 rx, u8 reg) -> u8;
-		auto indirect(u8 rx, u8 base) -> u8;
-		auto indirect_disp_8(u8 rx, u8 base) -> u8;
-		auto indirect_disp_32(u8 rx, u8 base) -> u8;
 	} // namespace detail
 	
 	class assembler_backend {
@@ -57,50 +39,6 @@ namespace baremetal {
 		auto find_optimized_instruction(u32 index, const operand* operands) -> const instruction*;
 		static auto is_legal_variant(u32 a, u32 b, u8 imm_index) -> bool;
 		static auto is_legal_variant_unknown(u32 a, u32 b, u8 imm_index) -> bool;
-		void emit_instruction_prefix(const instruction* inst, const operand* operands);
-		void instruction_begin(const operand* operands);
-
-		// instruction opcodes
-		void emit_instruction_opcode(const instruction* inst, const operand* operands);
-		void emit_instruction_opcode_prefix(const instruction* inst, const operand* operands);
-
-		// REX opcode prefix
-		void emit_opcode_prefix_rex(const instruction* inst, const operand* operands);
-		void emit_opcode_prefix_rex_mem(const mem& memory);
-
-		// pretty much all other prefixes also need a REX prefix in some shape or form, these utility
-		// functions are used for generating them
-		auto get_instruction_rex(const instruction* inst, const operand* operands) -> u8;
-
-		// VEX opcode prefix
-		void emit_opcode_prefix_vex_three(const instruction* inst, const operand* operands);
-		void emit_opcode_prefix_vex_two(const instruction* inst, const operand* operands);
-		void emit_opcode_prefix_vex(const instruction* inst, const operand* operands);
-
-		// EVEX opcode prefix
-		void emit_opcode_prefix_evex(const instruction* inst, const operand* operands);
-
-		// instruction MOD/RM
-		void emit_instruction_mod_rm(const instruction* inst, const operand* operands);
-
-		// instruction SIB
-		void emit_instruction_sib(const operand* operands);
-		auto has_sib_byte(const operand* operands) -> bool;
-	
-		// instruction operands
-		void emit_operands(const instruction* inst, const operand* operands);
-		void emit_data_operand(u64 data, u16 bit_width);
-		auto get_current_inst_size() const -> u8;
-
-		// instruction prefix helpers
-		auto get_instruction_vvvv(const instruction* inst, const operand* operands) -> u8;
-		auto get_instruction_v(const instruction* inst, const operand* operands) -> u8;
-		auto get_mask_register(const instruction* inst, const operand* operands) -> u8;
-		auto get_instruction_map_select(const instruction* inst) -> u8;
-		auto get_evex_operand_type(const instruction* inst) -> u8;
-		auto get_instruction_imp(const instruction* inst) -> u8;
-		auto get_instruction_l(const instruction* inst) -> bool;
-		auto get_evex_zero(const instruction* inst) -> bool;
 	private:
 		u64 m_current_inst_begin;
 		u8 m_reg_count;
